% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynareR-package.R
\docType{package}
\name{DynareR-package}
\alias{DynareR}
\alias{DynareR-package}
\title{DynareR: A Seamless Integration of 'R' and 'Dynare'}
\description{
It allows running 'Dynare' program from base R and R Markdown. 'Dynare' is a software platform for handling a wide class of economic models, in particular dynamic stochastic general equilibrium ('DSGE') and overlapping generations ('OLG') models. This package does not only integrate R and Dynare but also serves as a 'Dynare' Knit-Engine for 'knitr' package. The package requires 'Dynare' (<https://www.dynare.org/>) and 'Octave' (<https://www.gnu.org/software/octave/download.html>). Write all your 'Dynare' commands in R or R Markdown chunk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=DynareR}
  \item Report bugs at \url{https://github.com/sagirumati/DynareR/issues}
}

}
\author{
\strong{Maintainer}: Sagiru Mati \email{smati@smati.com.ng} (\href{https://orcid.org/0000-0003-1413-3974}{ORCID})

}
\keyword{documentation}
