% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tex.R
\name{input_tex}
\alias{input_tex}
\title{Include TeX file in R Markdown or Quarto document.}
\usage{
input_tex(path, start = NA, end = NA)
}
\arguments{
\item{path}{Object or a character string representing the path to the \code{TeX} file}

\item{start}{Numeric. The start line(s) of the \code{TeX} file to include.}

\item{end}{Numeric. The last line(s) of the \code{TeX} file to include.}
}
\value{
Set of \code{TeX} text
}
\description{
Use this function to include TeX file in R Markdown or Quarto document.
}
\examples{
library(DynareR)
\dontrun{

input_tex("DynareR/TeXFolder/olsTable.tex")
}
}
\seealso{
Other important functions: 
\code{\link{DynareR-package}},
\code{\link{add_matlab_path}()},
\code{\link{add_path}()},
\code{\link{eng_dynare}()},
\code{\link{import_log}()},
\code{\link{include_IRF}()},
\code{\link{run_dynare}()},
\code{\link{run_models}()},
\code{\link{set_dynare_version}()},
\code{\link{set_matlab_path}()},
\code{\link{set_octave_path}()},
\code{\link{write_dyn}()},
\code{\link{write_mod}()}
}
\concept{important functions}
\keyword{documentation}
