% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dynare_version.R
\name{set_dynare_version}
\alias{set_dynare_version}
\title{Set \code{Dynare} path}
\usage{
set_dynare_version(dynare_version="")
}
\arguments{
\item{dynare_version}{Path to the Dynare executable}
}
\value{
Character
}
\description{
Use this function to set \code{Dynare} path
}
\examples{
library(DynareR)
\dontrun{
set_dynare_version('C:/Program Files/Dynare 10/Dynare10.exe')
}
}
\seealso{
eng_dynare
}
\keyword{documentation}
