% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_matlab_path.R
\name{set_matlab_path}
\alias{set_matlab_path}
\title{Set \code{Matlab} path}
\usage{
set_matlab_path(matlab_path = "matlab")
}
\arguments{
\item{matlab_path}{Path to the Matlab executable. This overwrites \code{set_octave_path()} function.}
}
\value{
Character
}
\description{
Use this function to set \code{Matlab} path
}
\examples{
library(DynareR)
\dontrun{
set_matlab_path('C:/Program Files/MATLAB/R2024a/bin/matlab')
}
}
\seealso{
Other important functions: 
\code{\link{DynareR-package}},
\code{\link{add_matlab_path}()},
\code{\link{add_path}()},
\code{\link{eng_dynare}()},
\code{\link{import_log}()},
\code{\link{include_IRF}()},
\code{\link{input_tex}()},
\code{\link{run_dynare}()},
\code{\link{run_models}()},
\code{\link{set_dynare_version}()},
\code{\link{set_octave_path}()},
\code{\link{write_dyn}()},
\code{\link{write_mod}()}
}
\concept{important functions}
\keyword{documentation}
