% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_octave_path.R
\name{set_octave_path}
\alias{set_octave_path}
\title{Set \code{Octave} path}
\usage{
set_octave_path(octave_path = "octave")
}
\arguments{
\item{octave_path}{Path to the Octave executable}
}
\value{
Character
}
\description{
Use this function to set \code{Octave} path
}
\examples{
library(DynareR)
\dontrun{
set_octave_path('C:/Program Files/GNU Octave/Octave-6.4.0/mingw64/bin/octave20.exe')
}
}
\seealso{
Other important functions: 
\code{\link{DynareR-package}},
\code{\link{add_matlab_path}()},
\code{\link{add_path}()},
\code{\link{eng_dynare}()},
\code{\link{import_log}()},
\code{\link{include_IRF}()},
\code{\link{input_tex}()},
\code{\link{run_dynare}()},
\code{\link{run_models}()},
\code{\link{set_dynare_version}()},
\code{\link{set_matlab_path}()},
\code{\link{write_dyn}()},
\code{\link{write_mod}()}
}
\concept{important functions}
\keyword{documentation}
