% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_log.R
\name{import_log}
\alias{import_log}
\title{Import \code{dynare} log file as a list of R dataframes.}
\usage{
import_log(path=".",model="")
}
\arguments{
\item{path}{A character string for the path to the \code{dynare} log file.}

\item{model}{Object or a character string representing the name of the \code{Dynare} model file (\code{.mod} or \code{.dyn} extension)}
}
\value{
Set of \code{Dynare} (open-source software for DSGE modelling) outputs
}
\description{
Use \code{write_dyn(model,code)}  if you want the \code{Dynare} file to live in the current working directory.
Use \code{write_dyn(model,code,path)}  if you want the \code{Dynare} file to live in the path different from the current working directory.
}
\examples{
library(DynareR)
\dontrun{
import_log(model="BKK")

# Alternatively, use the path to the log file

import_log(path="BKK/BKK.log")
}
}
\seealso{
Other important functions: 
\code{\link{DynareR}},
\code{\link{add_path}()},
\code{\link{eng_dynare}()},
\code{\link{include_IRF}()},
\code{\link{run_dynare}()},
\code{\link{run_models}()},
\code{\link{set_dynare_version}()},
\code{\link{set_octave_path}()},
\code{\link{write_dyn}()},
\code{\link{write_mod}()}
}
\concept{important functions}
\keyword{documentation}
