% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_IRF.R
\name{include_IRF}
\alias{include_IRF}
\title{Embed the graphs Impulse Response Function (IRF) in R Markdown document}
\usage{
include_IRF(model="",IRF="",path="")
}
\arguments{
\item{model}{Object or a character string representing the name of the model excluding \code{.mod} or \code{.dyn} file extension}

\item{IRF}{A character string for the name of the Impulse Response Function as defined in the \code{Dynare} codes.}

\item{path}{Object or a character string representing the path to the folder for the \code{Dynare} file. The current working directory is the default \code{path}. Specify the \code{path} only if the \code{Dynare} files live in different path from the current working directory.}
}
\value{
Set of \code{Dynare} (open-source software for DSGE modelling) outputs
}
\description{
Use this function to include \code{Dynare} IRF into the R Markdown document
}
\examples{
knitr::knit_engines$set(dynare = DynareR::eng_dynare)
library(DynareR)
}
\references{
Bob Rudis (2015).Running Go language chunks in R Markdown (Rmd) files. Available at:  https://gist.github.com/hrbrmstr/9accf90e63d852337cb7

Yihui Xie (2019). knitr: A General-Purpose Package for Dynamic Report Generation in R. R package version 1.24.

Yihui Xie (2015) Dynamic Documents with R and knitr. 2nd edition. Chapman and Hall/CRC. ISBN 978-1498716963

Yihui Xie (2014) knitr: A Comprehensive Tool for Reproducible Research in R. In Victoria Stodden, Friedrich Leisch and Roger D. Peng, editors, Implementing Reproducible Computational Research. Chapman and Hall/CRC. ISBN 978-1466561595
}
\seealso{
write_mod write_dynare run_model run_dynare
}
\author{
Sagiru Mati, ORCID: 0000-0003-1413-3974, https://smati.com.ng
\itemize{
\item Yusuf Maitama Sule (Northwest) University Kano, Nigeria
\item SMATI Academy
}
}
\keyword{documentation}
