% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{run_models}
\alias{run_models}
\title{Run multiple \bold{existing} \code{mod} or \code{dyn} files.}
\usage{
run_models(model = "*", import_log = FALSE)
}
\arguments{
\item{model}{Object or a  vector of character strings representing the names of the \code{Dynare} model files excluding \code{.mod} or \code{.dyn} file extension}

\item{import_log}{Logical. Whether or not to import \code{dynare} log file.}
}
\value{
Set of \code{Dynare} (open-source software for DSGE modelling) outputs
}
\description{
Use this function to execute multiple \bold{existing} \code{Dynare} files.
Use \code{run_models(file)}  if the Dynare files live in the current working directory.
Use \code{run_models(file,path)}  if the Dynare files live in the path different from the current working directory.
}
\examples{
library(DynareR)

# Provide the list of the `Dynare` files in a vector

# Ensure that "example1.mod","example2.mod","agtrend.mod"
# and "bkk.mod" live in the path.

\dontrun{

run_models(model=c("example1","example2","agtrend","bkk"))

# You can run all models that live in "DynareR/run_models/" folder

run_models(path="DynareR/run_models")
}
}
\seealso{
Other important functions: 
\code{\link{DynareR}},
\code{\link{add_path}()},
\code{\link{eng_dynare}()},
\code{\link{import_log}()},
\code{\link{include_IRF}()},
\code{\link{run_dynare}()},
\code{\link{set_dynare_version}()},
\code{\link{set_octave_path}()},
\code{\link{write_dyn}()},
\code{\link{write_mod}()}
}
\concept{important functions}
\keyword{documentation}
