% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dynare_version.R
\name{set_dynare_version}
\alias{set_dynare_version}
\title{Set \code{Dynare} version}
\usage{
set_dynare_version(dynare_version="")
}
\arguments{
\item{dynare_version}{Character representing \code{Dynare} version (for example 6.1, 4.6.1 and so on). This has effect on Windows only.}
}
\value{
Character
}
\description{
Use this function to set \code{Dynare} version
}
\examples{
library(DynareR)
\dontrun{

# If you want to use the development version of Dynare

set_dynare_version("6-unstable-2022-04-03-0800-700a0e3a") # The development version of Dynare

# If you want to use Dynare version 5.2

set_dynare_version("5.2")
}
}
\seealso{
Other important functions: 
\code{\link{DynareR}},
\code{\link{add_matlab_path}()},
\code{\link{add_path}()},
\code{\link{eng_dynare}()},
\code{\link{import_log}()},
\code{\link{include_IRF}()},
\code{\link{input_tex}()},
\code{\link{run_dynare}()},
\code{\link{run_models}()},
\code{\link{set_octave_path}()},
\code{\link{write_dyn}()},
\code{\link{write_mod}()}
}
\concept{important functions}
\keyword{documentation}
