% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dynare_version.R
\name{set_dynare_version}
\alias{set_dynare_version}
\title{Set \code{Dynare} path}
\usage{
set_dynare_version(dynare_version="")
}
\arguments{
\item{dynare_version}{Path to the Dynare executable}
}
\value{
Character
}
\description{
Use this function to set \code{Dynare} path
}
\examples{
library(DynareR)
\dontrun{

# If you want to use the development version of Dynare

set_dynare_version("6-unstable-2022-04-03-0800-700a0e3a") # The development version of Dynare

# If you want to use Dynare version 5.2

set_dynare_version("5.2")
}
}
\seealso{
eng_dynare
}
\keyword{documentation}
