% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_octave_path.R
\name{set_octave_path}
\alias{set_octave_path}
\title{Set \code{Octave} path}
\usage{
set_octave_path(octave_system_path="octave")
}
\arguments{
\item{octave_system_path}{Path to the Octave executable}
}
\value{
Character
}
\description{
Use this function to set \code{Octave} path
}
\examples{
library(DynareR)
\dontrun{
set_octave_path('C:/Program Files/GNU Octave/Octave-6.4.0/mingw64/bin')
}
}
\seealso{
eng_dynare
}
\keyword{documentation}
